refactor(redux): migrate map and layers
