+++ src/features/beacons/actions.js
@@ -1,5 +1,5 @@
 import { beaconIdToGlobalId } from '~/model/identifiers';
-import { setSelectedFeatures } from '~/reducers/map/selection';
+import { setSelectedFeatures } from '~/features/map/selection';
 
 /**
  * Action factory that creates an action that sets the set of selected
