+++ src/features/weather/saga.js
@@ -2,7 +2,7 @@ import { delay, put, putResolve, race, select, take } from 'redux-saga/effects';
 
 import { isConnected } from '~/features/servers/selectors';
 import { setCurrentServerConnectionState } from '~/features/servers/slice';
-import { updateMapViewSettings } from '~/reducers/map/view';
+import { updateMapViewSettings } from '~/features/map/view';
 import { getMapViewCenterPosition } from '~/selectors/map';
 
 import { updateWeatherData } from './actions';
