+++ src/model/mutations.js
@@ -5,7 +5,7 @@ import {
   moveOutdoorShowOriginByMapCoordinateDelta,
   rotateOutdoorShowOrientationByAngle,
 } from '~/features/show/actions';
-import { updateFlatEarthCoordinateSystem } from '~/reducers/map/origin';
+import { updateFlatEarthCoordinateSystem } from '~/features/map/origin';
 import { toDegrees } from '~/utils/math';
 
 import { createFeatureFromOpenLayers } from './features';
