+++ src/views/layers/LayerList.jsx
@@ -12,7 +12,7 @@ import ListItemText from '@material-ui/core/ListItemText';
 import AddCircleOutline from '@material-ui/icons/AddCircleOutline';
 
 import { showLayerSettingsDialog } from '~/reducers/dialogs/layer-settings.js';
-import { addLayer } from '~/actions/layers';
+import { addLayer } from '~/features/map/layers';
 import { selectableListOf } from '~/components/helpers/lists';
 import { labelForLayerType, iconForLayerType } from '~/model/layers';
 import { getLayersInTopmostFirstOrder } from '~/selectors/ordered';
