+++ src/views/map/layers/heatmap.jsx
@@ -22,7 +22,7 @@ import Switch from '@material-ui/core/Switch';
 import Select from '@material-ui/core/Select';
 import TextField from '@material-ui/core/TextField';
 
-import { setLayerParametersById } from '~/actions/layers';
+import { setLayerParametersById } from '~/features/map/layers';
 import SubscriptionDialog from '~/components/dialogs/SubscriptionDialog';
 import messageHub from '~/message-hub';
 import HashedMap from '~/utils/hashedmap';
