+++ src/views/map/layers/uavtrace.jsx
@@ -10,7 +10,7 @@ import { layer as olLayer, source } from '@collmot/ol-react';
 
 import TextField from '@material-ui/core/TextField';
 
-import { setLayerParametersById } from '~/actions/layers';
+import { setLayerParametersById } from '~/features/map/layers';
 import SwatchesColorPicker from '~/components/SwatchesColorPicker';
 import flock from '~/flock';
 import { mapViewCoordinateFromLonLat } from '~/utils/geography';
