+++ src/views/map/layers/untyped.jsx
@@ -8,7 +8,7 @@ import CardHeader from '@material-ui/core/CardHeader';
 import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 
-import { changeLayerType } from '~/actions/layers';
+import { changeLayerType } from '~/features/map/layers';
 import {
   areMultipleInstancesAllowedForLayerType,
   LayerTypes,
